test(DMI_equations): Move DMI_equations unit tests to JSON format.
